+++ private/AdfObject.class.ps1
@@ -51,4 +51,6 @@ class AdfObject {
 
 }
 
-Set-Variable ADF_FOLDERS -option ReadOnly -value ('integrationRuntime', 'pipeline', 'dataset', 'dataflow', 'linkedService', 'trigger')
+if (!(Get-Variable ADF_FOLDERS -ErrorAction:SilentlyContinue)) {
+    Set-Variable ADF_FOLDERS -option ReadOnly -value ('integrationRuntime', 'pipeline', 'dataset', 'dataflow', 'linkedService', 'trigger')
+}