+++ changelog.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Removed workaround for deployment of Azure Integration Runtimes.
 ### Added
 * List of filtering rules (Includes/Excludes) can be provided by file
+* Config file allows defining a property to be added or removed
 
 ## [0.13.0] - 2020-07-08
 ### Fixed